Create codonUsageBias
This script compared the codon usage bias for a list of aligned nucleotide sequences ("seqstring") grouped according to a matrix ("cluster").  The script uses analysis of variance to determine whether groups display similar bias for a particular codon or not.  The number of codons displaying similar bias are summed and the results are generated as a comparison matrix displaying the number of shared codon biases between groups.